Add support for using Docker (#23)
Through Docker we can offload the complexity of the setup process. It enables users and developers to have a local website running fairly quickly. It also enables developers to add more complicated dependencies without having to worry about non-technical users, as the complexity is captured by Docker.

The build workflow uses the same image, which guarantees that if it works on your computer then it will also work and compile in the CI pipeline.